requirements.txt
langchain==0.3.14
langchain-community==0.3.14
langchain-text-splitters==0.3.4
chromadb==0.5.23
sentence-transformers==3.3.1
google-generativeai==0.8.4
fastapi==0.115.6
uvicorn==0.34.0
streamlit==1.41.1
python-dotenv==1.0.1
pypdf==5.1.0
python-docx==1.1.2
requests==2.32.3
beautifulsoup4==4.12.3